move localizer + moveonglobe code clean-up
