+++ services/motion/localizer.go
@@ -1,10 +1,11 @@
-// Package localizer introduces an interface which both slam and movementsensor can satisfy when wrapped respectively
-package localizer
+package motion
 
 import (
 	"context"
+	"fmt"
 
 	"go.viam.com/rdk/components/movementsensor"
+	"go.viam.com/rdk/resource"
 	"go.viam.com/rdk/services/slam"
 	"go.viam.com/rdk/spatialmath"
 )
@@ -12,6 +13,18 @@ import (
 // Localizer is an interface which both slam and movementsensor can satisfy when wrapped respectively.
 type Localizer interface {
 	GlobalPosition(context.Context) (spatialmath.Pose, error)
+}
+
+// NewLocalizer constructs either a SLAMLocalizer or MovementSensorLocalizer from the given resource
+func NewLocalizer(ctx context.Context, res resource.Resource) (Localizer, error) {
+	switch res := res.(type) {
+	case slam.Service:
+		return &SLAMLocalizer{Service: res}, nil
+	case movementsensor.MovementSensor:
+		return &MovementSensorLocalizer{MovementSensor: res}, nil
+	default:
+		return nil, fmt.Errorf("cannot localize on resource of type %T", res)
+	}
 }
 
 // SLAMLocalizer is a struct which only wraps an existing slam service.
